# docs/swagger.yaml
swagger: "2.0"
info:
  description: "This is the documentation for the NIMPI project"
  version: "1.0.0"
  title: "NIMPI"
basePath: "/v1"
...
schemes:
- "https"
- "http"
paths:
...
  /document/{id}:
    get:
      tags:
      - "Document"
      summary: "Finds document by term"
      description: ""
      operationId: "findDocumentsByID"
      produces:
      - "application/json"
      parameters:
        - in: path
          name: id
          required: true
          type: integer
          description: Find by id
      responses:
        '200':
          description: "successful operation"
          schema:
            $ref: "#/definitions/Document"
        '401':
          description: Unauthorized query, check the connection requirements.
        '404':
          description: A document with this id was not found.
        '500':
          description: Internal Error Server.
    put:
      tags:
      - "Document"
      summary: "Update an existing document"
      description: ""
      operationId: "updateDocument"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: id
        in: "path"
        description: "ID of document to return"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "body"
        description: "Update a document's metadata"
        required: true
        schema:
          $ref: "#/definitions/Document"
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized request, check connection parameters.
        '404':
          description: A document with this id was not found.
        '500':
          description: Internal Error Server.
      security:
      - admin_auth:
        - "write:document"
        - "read:document"
    delete:
      tags:
      - "Document"
      summary: "Deletes a Document"
      description: ""
      operationId: "deleteDocument"
      produces:
      - "application/json"
      parameters:
      - name: id
        in: "path"
        description: "ID of document to return"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "body"
        description: "List of user object"
        required: true
        schema:
          type: "array"
          items:
            $ref: "#/definitions/User"
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized request, check connection parameters.
        '404':
          description: A document with this id was not found.
        '500':
          description: Internal Error Server.
      security:
      - admin_auth:
        - "write:document"
        - "read:document"
...
securityDefinitions:
  admin_auth:
    type: "oauth2"
    authorizationUrl: "v1/oauth/dialog"
    flow: "implicit"
    scopes:
      write:document: "can make changes to the system, read permission"
      read:document: "can just view, can't write"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"   
definitions:
  Document:
    type: "object"
    properties:
      id:
        type: "integer"
      title:
        type: "string"
        xml:
          name: "tag"
          wrapped: true
      description:
        type: "string"
      date:
        type: "string"
        format: "date-time"
      year:
        type: "integer"
      publisherId:
        type: "integer"
      type:
        type: "string"
      key:
        type: "string"
      size:
        type: "float"
      path:
        type: "string"
      dateCreated:
        type: "string"
        format: "date-time"
      lastModification:
        type: "string"
        format: "date-time"
      tags:
        type: "string"
      userId:
        type: "integer"
...
  User:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
      enable:
        type: "boolean"
        default: true
...